moved the code from tms classes to tm classes (was a mistake to write these in tms classes)
